+++ packages/orm/angel_orm_generator/lib/src/orm_build_context.dart
@@ -74,7 +74,7 @@ Future<OrmBuildContext?> buildOrmContext(
               .firstAnnotationOf(clazz)) !=
       null) {
     if (clazz.supertype != null) {
-      clazz = clazz.supertype!.element2;
+      clazz = clazz.supertype!.element;
     }
   }
 
@@ -172,7 +172,7 @@ Future<OrmBuildContext?> buildOrmContext(
             }
 
             var modelType = firstModelAncestor(refType) ?? refType;
-            var modelTypeElement = modelType.element2;
+            var modelTypeElement = modelType.element;
 
             if (modelTypeElement != null) {
               foreign = await buildOrmContext(
@@ -188,7 +188,7 @@ Future<OrmBuildContext?> buildOrmContext(
               if (through != null && through is InterfaceType) {
                 throughContext = await buildOrmContext(
                     cache,
-                    through.element2,
+                    through.element,
                     ConstantReader(const TypeChecker.fromRuntime(Serializable)
                         .firstAnnotationOf(modelTypeElement)),
                     buildStep,
@@ -253,9 +253,8 @@ Future<OrmBuildContext?> buildOrmContext(
         // Unfortunately, the analyzer library provides little to nothing
         // in the way of reading enums from source, so here's a hack.
         var joinTypeType = (joinTypeRdr.type as InterfaceType);
-        var enumFields = joinTypeType.element2.fields
-            .where((f) => f.isEnumConstant)
-            .toList();
+        var enumFields =
+            joinTypeType.element.fields.where((f) => f.isEnumConstant).toList();
 
         for (var i = 0; i < enumFields.length; i++) {
           if (enumFields[i].computeConstantValue() == joinTypeRdr) {
@@ -295,7 +294,7 @@ Future<OrmBuildContext?> buildOrmContext(
             if (foreign != null) {
               if (isSpecialId(foreign, foreignField)) {
                 // Use integer
-                type = field.type.element2?.library?.typeProvider.intType
+                type = field.type.element?.library?.typeProvider.intType
                     as DartType;
 
                 //type = field.type.element?.context.typeProvider.intType;
@@ -364,7 +363,7 @@ ColumnType inferColumnType(DartType type) {
   if (const TypeChecker.fromRuntime(List).isAssignableFromType(type)) {
     return ColumnType.jsonb;
   }
-  if (type is InterfaceType && type.element2 is EnumElement) {
+  if (type is InterfaceType && type.element is EnumElement) {
     return ColumnType.int;
   }