+++ packages/orm/angel_orm_generator/lib/src/orm_generator.dart
@@ -271,14 +271,14 @@ class OrmGenerator extends GeneratorForAnnotation<Orm> {
                 //    .call([expr.property('toString').call([])]);
                 expr = refer('mapToDouble').call([expr]);
               } else if (fType is InterfaceType &&
-                  fType.element2 is EnumElement) {
+                  fType.element is EnumElement) {
                 var isNull = expr.equalTo(literalNull);
                 expr = isNull.conditional(literalNull,
                     type.property('values').index(expr.asA(refer('int'))));
               } else if (fType.isDartCoreBool) {
                 // Generated Code: mapToBool(row[i])
                 expr = refer('mapToBool').call([expr]);
-              } else if (fType.element2?.displayName == 'DateTime') {
+              } else if (fType.element?.displayName == 'DateTime') {
                 // Generated Code: mapToDateTime(row[i])
                 if (fType.nullabilitySuffix == NullabilitySuffix.question) {
                   expr = refer('mapToNullableDateTime').call([expr]);
@@ -300,7 +300,7 @@ class OrmGenerator extends GeneratorForAnnotation<Orm> {
                   defaultRef = CodeExpression(Code('0.0'));
                 } else if (fType.isDartCoreInt || fType.isDartCoreNum) {
                   defaultRef = CodeExpression(Code('0'));
-                } else if (fType.element2?.displayName == 'DateTime') {
+                } else if (fType.element?.displayName == 'DateTime') {
                   defaultRef = CodeExpression(
                       Code('DateTime.parse("1970-01-01 00:00:00")'));
                 } else if (fType.isDartCoreList) {
@@ -748,7 +748,7 @@ class OrmGenerator extends GeneratorForAnnotation<Orm> {
           builderType = TypeReference((b) => b
             ..symbol = 'NumericSqlExpressionBuilder'
             ..types.add(refer(typeName)));
-        } else if (type is InterfaceType && type.element2 is EnumElement) {
+        } else if (type is InterfaceType && type.element is EnumElement) {
           builderType = TypeReference((b) => b
             ..symbol = 'EnumSqlExpressionBuilder'
             ..types.add(convertTypeReference(type)));
@@ -884,7 +884,7 @@ class OrmGenerator extends GeneratorForAnnotation<Orm> {
         clazz.methods.add(Method((b) {
           var value = refer('values').index(literalString(name!));
 
-          if (fType is InterfaceType && fType.element2 is EnumElement) {
+          if (fType is InterfaceType && fType.element is EnumElement) {
             var asInt = value.asA(refer('int'));
             var t = convertTypeReference(fType);
             value = t.property('values').index(asInt);
@@ -918,7 +918,7 @@ class OrmGenerator extends GeneratorForAnnotation<Orm> {
         clazz.methods.add(Method((b) {
           Expression value = refer('value');
 
-          if (fType is InterfaceType && fType.element2 is EnumElement) {
+          if (fType is InterfaceType && fType.element is EnumElement) {
             value = CodeExpression(Code('value?.index'));
           } else if (const TypeChecker.fromRuntime(List)
               .isAssignableFromType(fType)) {